requirements.txt
# ExtendBot - Extended Exchange Trading Bot
# Python 3.10+ required

# Extended Exchange SDK
# Install from GitHub: pip install git+https://github.com/x10xchange/python_sdk.git@starknet
x10-python-trading @ git+https://github.com/x10xchange/python_sdk.git@starknet

# Async HTTP client
aiohttp>=3.9.0
httpx>=0.25.0

# WebSocket support
websockets>=12.0

# Environment variables
python-dotenv>=1.0.0

# Telegram bot
python-telegram-bot>=20.7

# Database (PostgreSQL)
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data analysis
pandas>=2.1.0
numpy>=1.26.0

# Technical analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Machine Learning (optional)
scikit-learn>=1.3.0
joblib>=1.3.0

# Decimal/Math utilities
python-decimal>=1.3

# Date/Time utilities
python-dateutil>=2.8.2

# Logging
coloredlogs>=15.0.1

# Rate limiting
aiolimiter>=1.1.0

# Crypto/Stark curve
starknet-py>=0.20.0
starknet-crypto>=0.1.0

# Utilities
pydantic>=2.5.0
orjson>=3.9.0
